+++ my-portfolio/public/data/certificationsData.js
