+++ my-portfolio/public/data/educationData.js
