+++ my-portfolio/public/data/goalsData.js
