+++ my-portfolio/public/data/heroData.js
