+++ my-portfolio/public/data/projectsData.js
