+++ my-portfolio/public/data/socialsData.js
