+++ my-portfolio/src/mobileComponents/sections/DigitalPresence/index.jsx
@@ -3,7 +3,7 @@ import {
   Copy, Check, ExternalLink, ArrowRight,
   Sparkles, Globe, Mail, Phone
 } from 'lucide-react';
-import { contactConfig, contactSources, quickLinks } from '../../../data/contactData';
+import { contactConfig, contactSources, quickLinks } from '../../../../public/data/contactData';
 
 const ContactButton = ({ link, index }) => (
   <a
