+++ my-portfolio/src/mobileComponents/sections/Testimonials/components/TestimonialCard.jsx
@@ -2,7 +2,7 @@
 import React, { useState, useRef } from 'react';
 import { Quote, Star, ChevronUp } from 'lucide-react';
 import { motion, AnimatePresence } from 'framer-motion';
-import { testimonials } from '../../../../data/testimonialsData'; // Add this import
+import { testimonials } from '../../../../../public/data/testimonialsData'; // Add this import
 
 
 const MobileTestimonialCard = ({ testimonial, isExpanded, onToggle }) => (
